Move class to file
